+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_ebo.cpp
@@ -992,7 +992,7 @@ bool AArch64Ebo::CombineMultiplySub(Insn *insn, OpndInfo *opndInfo, bool is64bit
 
 bool CheckInsnRefField(Insn &insn, size_t opndIndex) {
   if (insn.IsAccessRefField() && static_cast<AArch64Insn&>(insn).AccessMem()) {
-    Operand &opnd0 = insn.GetOperand(opndIndex);
+    Operand &opnd0 = insn.GetOperand(static_cast<uint32>(opndIndex));
     if (opnd0.IsRegister()) {
       return true;
     }
