+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_emitter.cpp
@@ -380,7 +380,7 @@ void AArch64AsmEmitter::RecordRegInfo(FuncEmitInfo &funcEmitInfo) {
           if (regType == kRegTyCc || regType == kRegTyVary) {
             continue;
           }
-          bool isDef = md->GetOperand(static_cast<int>(i))->IsRegDef();
+          bool isDef = md->GetOperand(i)->IsRegDef();
           if (isDef) {
             referedRegs.insert(static_cast<RegOperand&>(opnd).GetRegisterNumber());
           }
@@ -2054,10 +2054,10 @@ struct CfiDescr {
 static CfiDescr cfiDescrTable[cfi::kOpCfiLast + 1] = {
 #define CFI_DEFINE(k, sub, n, o0, o1, o2) \
   { ".cfi_" #k, n, { Operand::kOpd##o0, Operand::kOpd##o1, Operand::kOpd##o2 } },
-  #define ARM_DIRECTIVES_DEFINE(k, sub, n, o0, o1, o2) \
+#define ARM_DIRECTIVES_DEFINE(k, sub, n, o0, o1, o2) \
   { "." #k, n, { Operand::kOpd##o0, Operand::kOpd##o1, Operand::kOpd##o2 } },
-  #include "cfi.def"
-  #undef CFI_DEFINE
+#include "cfi.def"
+#undef CFI_DEFINE
 #undef ARM_DIRECTIVES_DEFINE
   { ".cfi_undef", 0, { Operand::kOpdUndef, Operand::kOpdUndef, Operand::kOpdUndef } }
 };
@@ -2106,7 +2106,7 @@ void AArch64AsmEmitter::EmitAArch64DbgInsn(Emitter &emitter, const Insn &insn) {
 
 bool AArch64AsmEmitter::CheckInsnRefField(Insn &insn, size_t opndIndex) const {
   if (insn.IsAccessRefField() && static_cast<AArch64Insn&>(insn).AccessMem()) {
-    Operand &opnd0 = insn.GetOperand(opndIndex);
+    Operand &opnd0 = insn.GetOperand(static_cast<uint32>(opndIndex));
     if (opnd0.IsRegister()) {
       static_cast<RegOperand&>(opnd0).SetRefField(true);
       return true;
